requirements.txt
Django==4.1.6
django-extensions==3.2.1
django-debug-toolbar==3.8.1
django-admin-rangefilter
django-import-export
ipython



# flake8
flake8
flake8-builtins
flake8-print
